Changes for ZeroMQ libzmq 3.1.0 release
Changes made to the hxzmq wrapper classes, mostly in the ZMQ.hx class,
for the changes made in the ZeroMQ 3.1.0 release.

The changes are not backwardly compatible with the previous 2.1.x
libzmq version; this new version of hxzmq will only work alongside a
3.1.x version libzmq native library.

See http://www.zeromq.org/docs:3-1-upgrade for details of the changes
made in the libzmq library, and therefore in hxzmq.